move video-, depth-, disparity-frame into own header
