+++ src/archive.cpp
@@ -5,6 +5,7 @@
 #include "frame-archive.h"
 #include "core/pose-frame.h"
 #include "core/motion-frame.h"
+#include "core/disparity-frame.h"
 
 
 namespace librealsense
