+++ src/core/CMakeLists.txt
@@ -6,10 +6,13 @@ target_sources(${LRS_TARGET}
         "${CMAKE_CURRENT_LIST_DIR}/streaming.h"
         "${CMAKE_CURRENT_LIST_DIR}/debug.h"
         "${CMAKE_CURRENT_LIST_DIR}/advanced_mode.h"
+        "${CMAKE_CURRENT_LIST_DIR}/depth-frame.h"
+        "${CMAKE_CURRENT_LIST_DIR}/disparity-frame.h"
         "${CMAKE_CURRENT_LIST_DIR}/roi.h"
         "${CMAKE_CURRENT_LIST_DIR}/motion.h"
         "${CMAKE_CURRENT_LIST_DIR}/motion-frame.h"
         "${CMAKE_CURRENT_LIST_DIR}/video.h"
+        "${CMAKE_CURRENT_LIST_DIR}/video-frame.h"
         "${CMAKE_CURRENT_LIST_DIR}/options.h"
         "${CMAKE_CURRENT_LIST_DIR}/info.h"
         "${CMAKE_CURRENT_LIST_DIR}/extension.h"
