+++ src/frame.h
@@ -297,179 +297,5 @@ private:
     std::shared_ptr< stream_profile_interface > stream;
 };
 
-class video_frame : public frame
-{
-public:
-    video_frame()
-        : frame()
-        , _width( 0 )
-        , _height( 0 )
-        , _bpp( 0 )
-        , _stride( 0 )
-    {
-    }
-
-    int get_width() const { return _width; }
-    int get_height() const { return _height; }
-    int get_stride() const { return _stride; }
-    int get_bpp() const { return _bpp; }
-
-    void assign( int width, int height, int stride, int bpp )
-    {
-        _width = width;
-        _height = height;
-        _stride = stride;
-        _bpp = bpp;
-    }
-
-private:
-    int _width, _height, _bpp, _stride;
-};
-
-MAP_EXTENSION( RS2_EXTENSION_VIDEO_FRAME, librealsense::video_frame );
-
-class depth_frame : public video_frame
-{
-public:
-    depth_frame()
-        : video_frame()
-    {
-    }
-
-    frame_interface * publish( std::shared_ptr< archive_interface > new_owner ) override
-    {
-        return video_frame::publish( new_owner );
-    }
-
-    void keep() override
-    {
-        if( _original )
-            _original->keep();
-        video_frame::keep();
-    }
-
-    float get_distance( int x, int y ) const;
-
-    float get_units() const { return additional_data.depth_units; }
-
-    void set_original( frame_holder h )
-    {
-        _original = std::move( h );
-        attach_continuation( frame_continuation(
-            [this]() {
-                if( _original )
-                {
-                    _original = {};
-                }
-            },
-            nullptr ) );
-    }
-
-protected:
-    static float query_units( const std::shared_ptr< sensor_interface > & sensor )
-    {
-        if( sensor != nullptr )
-        {
-            try
-            {
-                auto depth_sensor = As< librealsense::depth_sensor >( sensor );
-                if( depth_sensor != nullptr )
-                {
-                    return depth_sensor->get_depth_scale();
-                }
-                else
-                {
-                    // For playback sensors
-                    auto extendable = As< librealsense::extendable_interface >( sensor );
-                    if( extendable
-                        && extendable->extend_to(
-                            TypeToExtension< librealsense::depth_sensor >::value,
-                            (void **)( &depth_sensor ) ) )
-                    {
-                        return depth_sensor->get_depth_scale();
-                    }
-                }
-            }
-            catch( const std::exception & e )
-            {
-                LOG_ERROR( "Failed to query depth units from sensor. " << e.what() );
-            }
-            catch( ... )
-            {
-                LOG_ERROR( "Failed to query depth units from sensor" );
-            }
-        }
-        else
-        {
-            LOG_WARNING( "sensor was nullptr" );
-        }
-
-        return 0;
-    }
-
-    frame_holder _original;
-};
-
-MAP_EXTENSION( RS2_EXTENSION_DEPTH_FRAME, librealsense::depth_frame );
-
-// Disparity frame provides an alternative representation of the depth data for stereo-based depth
-// sensors In addition for the depth frame API it allows to query the stereoscopic baseline required
-// to transform depth to disparity and vice versa
-class disparity_frame : public depth_frame
-{
-public:
-    disparity_frame()
-        : depth_frame()
-    {
-    }
-
-    // TODO Refactor to framemetadata
-    float get_stereo_baseline( void ) const { return query_stereo_baseline( this->get_sensor() ); }
-
-protected:
-    static float query_stereo_baseline( const std::shared_ptr< sensor_interface > & sensor )
-    {
-        if( sensor != nullptr )
-        {
-            try
-            {
-                auto stereo_sensor = As< librealsense::depth_stereo_sensor >( sensor );
-                if( stereo_sensor != nullptr )
-                {
-                    return stereo_sensor->get_stereo_baseline_mm();
-                }
-                else
-                {
-                    // For playback sensors
-                    auto extendable = As< librealsense::extendable_interface >( sensor );
-                    if( extendable
-                        && extendable->extend_to(
-                            TypeToExtension< librealsense::depth_stereo_sensor >::value,
-                            (void **)( &stereo_sensor ) ) )
-                    {
-                        return stereo_sensor->get_stereo_baseline_mm();
-                    }
-                }
-            }
-            catch( const std::exception & e )
-            {
-                LOG_ERROR( "Failed to query stereo baseline from sensor. " << e.what() );
-            }
-            catch( ... )
-            {
-                LOG_ERROR( "Failed to query stereo baseline from sensor" );
-            }
-        }
-        else
-        {
-            LOG_WARNING( "sensor was nullptr" );
-        }
-
-        return 0;
-    }
-};
-
-MAP_EXTENSION( RS2_EXTENSION_DISPARITY_FRAME, librealsense::disparity_frame );
-
 
 }  // namespace librealsense
