+++ src/media/ros/ros_reader.cpp
@@ -16,6 +16,7 @@
 #include "std_msgs/Float32MultiArray.h"
 #include <src/core/pose-frame.h>
 #include <src/core/motion-frame.h>
+#include <src/core/video-frame.h>
 
 #include <rsutils/string/from.h>
