+++ src/points.cpp
@@ -2,6 +2,7 @@
 // Copyright(c) 2021 Intel Corporation. All Rights Reserved.
 #include "points.h"
 #include "core/video.h"
+#include "core/video-frame.h"
 #include <fstream>
 
 #define MIN_DISTANCE 1e-6
