+++ src/proc/disparity-transform.h
@@ -5,8 +5,9 @@
 
 #pragma once
 
-#include "../include/librealsense2/hpp/rs_frame.hpp"
-#include "../include/librealsense2/hpp/rs_processing.hpp"
+#include <librealsense2/hpp/rs_frame.hpp>
+#include <librealsense2/hpp/rs_processing.hpp>
+#include <src/core/depth-frame.h>
 #include "synthetic-stream.h"
 
 namespace librealsense
