+++ src/proc/synthetic-stream.cpp
@@ -5,6 +5,7 @@
 
 #include "core/video.h"
 #include "core/motion-frame.h"
+#include "core/depth-frame.h"
 #include "option.h"
 #include "context.h"
 #include "stream.h"
