+++ src/proc/threshold.cpp
@@ -1,9 +1,10 @@
 // License: Apache 2.0. See LICENSE file in root directory.
 // Copyright(c) 2017 Intel Corporation. All Rights Reserved.
 
-#include "../include/librealsense2/hpp/rs_sensor.hpp"
-#include "../include/librealsense2/hpp/rs_processing.hpp"
+#include <librealsense2/hpp/rs_sensor.hpp>
+#include <librealsense2/hpp/rs_processing.hpp>
 
+#include <src/core/depth-frame.h>
 #include "proc/synthetic-stream.h"
 #include "context.h"
 #include "environment.h"
