+++ src/rs.cpp
@@ -18,6 +18,7 @@
 #include "core/advanced_mode.h"
 #include "core/pose-frame.h"
 #include "core/motion-frame.h"
+#include "core/disparity-frame.h"
 #include "source.h"
 #include "core/processing.h"
 #include "proc/synthetic-stream.h"
