+++ src/software-device.cpp
@@ -3,6 +3,7 @@
 
 #include "software-device.h"
 #include "stream.h"
+#include "core/video-frame.h"
 
 #include <rsutils/string/from.h>
 #include <rsutils/deferred.h>
