+++ src/uvc-sensor.cpp
@@ -5,6 +5,7 @@
 #include "device.h"
 #include "stream.h"
 #include "global_timestamp_reader.h"
+#include "core/video-frame.h"
 
 
 namespace librealsense {
